+++ src/drop.test.ts
@@ -1,4 +1,4 @@
-import { createLazyInvocationCounter } from "../test/lazy_invocation_counter";
+import { createLazyInvocationCounter } from "../test/lazyInvocationCounter";
 import { drop } from "./drop";
 import { pipe } from "./pipe";
 import { take } from "./take";
