+++ src/flatten.test.ts
@@ -1,4 +1,4 @@
-import { createLazyInvocationCounter } from "../test/lazy_invocation_counter";
+import { createLazyInvocationCounter } from "../test/lazyInvocationCounter";
 import { find } from "./find";
 import { flatten } from "./flatten";
 import { pipe } from "./pipe";
