+++ src/isNumber.test.ts
@@ -2,7 +2,7 @@ import {
   ALL_TYPES_DATA_PROVIDER,
   TYPES_DATA_PROVIDER,
   type AllTypesDataProviderTypes,
-} from "../test/types_data_provider";
+} from "../test/typesDataProvider";
 import { isNumber } from "./isNumber";
 
 const dataFunction = (): string | 1 | 2 | 3 => 1;
