+++ src/isPlainObject.test.ts
@@ -3,7 +3,7 @@ import {
   TYPES_DATA_PROVIDER,
   TestClass,
   type AllTypesDataProviderTypes,
-} from "../test/types_data_provider";
+} from "../test/typesDataProvider";
 import { isPlainObject } from "./isPlainObject";
 
 describe("runtime", () => {
