+++ test/lazyInvocationCounter.ts
@@ -1,5 +1,6 @@
 import { map } from "../src/map";
 
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type, @typescript-eslint/explicit-module-boundary-types
 export const createLazyInvocationCounter = () => {
   const count = vi.fn();
   return {
